requirements.txt
predict-sdk
loguru
pyyaml
requests
python-dotenv
eth-account
web3
